Promote v2 pages to mainline
