+++ src/dashboard/src/pages/Job/Tabs/Brief.tsx
