+++ src/dashboard/src/pages/Job/Tabs/Endpoints.tsx
