+++ src/dashboard/src/pages/JobLegacy/Details/Endpoints.tsx
