+++ src/dashboard/src/pages/JobLegacy/Details/Log.tsx
