+++ src/dashboard/src/pages/JobLegacy/Details/Monitor.tsx
