+++ src/dashboard/src/pages/JobLegacy/Details/index.tsx
