+++ src/dashboard/src/pages/Jobs/JobsTable/DetailPanel.tsx
