+++ src/dashboard/src/pages/Jobs/JobsTable/columns/gpu.tsx
