+++ src/dashboard/src/pages/Jobs/JobsTable/columns/type.tsx
