+++ src/dashboard/src/pages/JobsLegacy/components/JobsOperationDialog.tsx
