+++ src/dashboard/src/pages/Submission/components/GPUCard.tsx
@@ -97,7 +97,7 @@ const ActionIconButton: React.FC<{cluster?: string}> = ({cluster}) => {
         open={open}
         onClose={onMenuClose}
       >
-        <MenuItem component={Link} to={"/cluster-status"}>Cluster Status</MenuItem>
+        <MenuItem component={Link} to={`/clusters/${cluster}`}>Cluster Status</MenuItem>
         <MenuItem component={Link} to={`/jobs/${cluster}`}>View Jobs</MenuItem>
       </Menu>
     </>
